small edits

+++ Assets/Scripts/FishingProcess/FishCatchScore.cs
@@ -5,7 +5,7 @@
 public class FishCatchScore
 {
     public FishSpecies species;
-    public float minScore;
-    public float maxScore;
+    public int minScore;
+    public int maxScore;
     public float timeAttached; // How long the fish is attached until it frees itself
 }
